+++ packages/agent-infra/mcp-servers/browser/src/index.ts
@@ -7,10 +7,10 @@
  * Copyright (c) Microsoft Corporation.
  * https://github.com/microsoft/playwright-mcp/blob/main/LICENSE
  */
+import { startSseAndStreamableHttpMcpServer } from 'mcp-http-server';
 import { program } from 'commander';
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
 import { createServer, getBrowser } from './server.js';
-import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 
 declare global {
   interface Window {
@@ -56,10 +56,10 @@ program
   // .option('--device <device>', 'device to emulate, for example: "iPhone 15"')
   .option('--executable-path <path>', 'path to the browser executable.')
   .option('--headless', 'run browser in headless mode, headed by default')
-  // .option(
-  //   '--host <host>',
-  //   'host to bind server to. Default is localhost. Use 0.0.0.0 to bind to all interfaces.',
-  // )
+  .option(
+    '--host <host>',
+    'host to bind server to. Default is localhost. Use 0.0.0.0 to bind to all interfaces.',
+  )
   // .option('--ignore-https-errors', 'ignore https errors')
   // .option(
   //   '--isolated',
@@ -71,7 +71,7 @@ program
   //   'disable the sandbox for all process types that are normally sandboxed.',
   // )
   // .option('--output-dir <path>', 'path to the directory for output files.')
-  // .option('--port <port>', 'port to listen on for SSE transport.')
+  .option('--port <port>', 'port to listen on for SSE and HTTP transport.')
   .option(
     '--proxy-bypass <bypass>',
     'comma-separated domains to bypass proxy, for example ".com,chromium.org,.domain.com"',
@@ -102,71 +102,86 @@ program
     try {
       console.log('[mcp-server-browser] options', options);
 
-      const server: McpServer = createServer({
-        ...((options.cdpEndpoint || options.wsEndpoint) && {
-          remoteOptions: {
-            wsEndpoint: options.wsEndpoint,
-            cdpEndpoint: options.cdpEndpoint,
-          },
-        }),
-        launchOptions: {
-          headless: options.headless,
-          executablePath: options.executablePath,
-          browserType: options.browser,
-          proxy: options.proxyServer,
-          proxyBypassList: options.proxyBypass,
-          args: [process.env.DISPLAY ? `--display=${process.env.DISPLAY}` : ''],
-          ...(options.viewportSize && {
-            defaultViewport: {
-              width: parseInt(options.viewportSize?.split(',')[0]),
-              height: parseInt(options.viewportSize?.split(',')[1]),
+      const createMcpServer = async () => {
+        const server = createServer({
+          ...((options.cdpEndpoint || options.wsEndpoint) && {
+            remoteOptions: {
+              wsEndpoint: options.wsEndpoint,
+              cdpEndpoint: options.cdpEndpoint,
             },
           }),
-          ...(options.userDataDir && {
-            userDataDir: options.userDataDir,
-          }),
-        },
-        contextOptions: {
-          userAgent: options.userAgent,
-        },
-        logger: {
-          info: (...args: any[]) => {
-            server.server.notification({
-              method: 'notifications/message',
-              params: {
-                level: 'warning',
-                logger: 'mcp-server-browser',
-                data: JSON.stringify(args),
+          launchOptions: {
+            headless: options.headless,
+            executablePath: options.executablePath,
+            browserType: options.browser,
+            proxy: options.proxyServer,
+            proxyBypassList: options.proxyBypass,
+            args: [
+              process.env.DISPLAY ? `--display=${process.env.DISPLAY}` : '',
+            ],
+            ...(options.viewportSize && {
+              defaultViewport: {
+                width: parseInt(options.viewportSize?.split(',')[0]),
+                height: parseInt(options.viewportSize?.split(',')[1]),
               },
-            });
+            }),
+            ...(options.userDataDir && {
+              userDataDir: options.userDataDir,
+            }),
+          },
+          contextOptions: {
+            userAgent: options.userAgent,
+          },
+          logger: {
+            info: (...args: any[]) => {
+              server.server.notification({
+                method: 'notifications/message',
+                params: {
+                  level: 'warning',
+                  logger: 'mcp-server-browser',
+                  data: JSON.stringify(args),
+                },
+              });
 
-            server.server.sendLoggingMessage({
-              level: 'info',
-              data: JSON.stringify(args),
-            });
+              server.server.sendLoggingMessage({
+                level: 'info',
+                data: JSON.stringify(args),
+              });
+            },
+            error: (...args: any[]) => {
+              server.server.sendLoggingMessage({
+                level: 'error',
+                data: JSON.stringify(args),
+              });
+            },
+            warn: (...args: any[]) => {
+              server.server.sendLoggingMessage({
+                level: 'warning',
+                data: JSON.stringify(args),
+              });
+            },
+            debug: (...args: any[]) => {
+              server.server.sendLoggingMessage({
+                level: 'debug',
+                data: JSON.stringify(args),
+              });
+            },
           },
-          error: (...args: any[]) => {
-            server.server.sendLoggingMessage({
-              level: 'error',
-              data: JSON.stringify(args),
-            });
-          },
-          warn: (...args: any[]) => {
-            server.server.sendLoggingMessage({
-              level: 'warning',
-              data: JSON.stringify(args),
-            });
-          },
-          debug: (...args: any[]) => {
-            server.server.sendLoggingMessage({
-              level: 'debug',
-              data: JSON.stringify(args),
-            });
-          },
-        },
-      });
-      const transport = new StdioServerTransport();
-      await server.connect(transport);
+        });
+
+        return server;
+      };
+      if (options.port || options.host) {
+        await startSseAndStreamableHttpMcpServer({
+          host: options.host,
+          port: options.port,
+          createMcpServer: async () => createMcpServer() as any,
+        });
+      } else {
+        const server = await createMcpServer();
+        const transport = new StdioServerTransport();
+        await server.connect(transport);
+      }
     } catch (error) {
       console.error('Error: ', error);
       process.exit(1);