+++ packages/agent-infra/mcp-client/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @agent-infra/mcp-client
 
+## 1.1.1
+
+### Patch Changes
+
+- a40b3c1: feat: add mcp http server
+- Updated dependencies [a40b3c1]
+  - @agent-infra/mcp-shared@1.1.1
+
+## 1.1.1-beta.3
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.1-beta.3
+
+## 1.1.1-beta.2
+
+### Patch Changes
+
+- feat: add mcp http server
+- Updated dependencies
+  - @agent-infra/mcp-shared@1.1.1-beta.2
+
+## 1.1.1-beta.1
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.1-beta.1
+
+## 1.1.1-beta.0
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.1-beta.0
+
 ## 1.1.0
 
 ### Minor Changes