+++ packages/agent-infra/mcp-servers/browser/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @agent-infra/mcp-server-browser
 
+## 1.1.1
+
+### Patch Changes
+
+- a40b3c1: feat: add mcp http server
+- 2598ea6: feat: native support sse and mcp serving by browser mcp
+- 82fe970: fix: hang up
+- 1e12e91: fix: schema empty not right
+- feat: support high performance sse and http server
+- Updated dependencies [a40b3c1]
+- Updated dependencies [82fe970]
+- Updated dependencies
+  - mcp-http-server@1.1.1
+
+## 1.1.1-beta.3
+
+### Patch Changes
+
+- fix: hang up
+- Updated dependencies
+  - mcp-http-server@1.1.1-beta.3
+
+## 1.1.1-beta.2
+
+### Patch Changes
+
+- feat: add mcp http server
+- Updated dependencies
+  - @agent-infra/mcp-http-server@1.1.1-beta.2
+
+## 1.1.1-beta.1
+
+### Patch Changes
+
+- fix: schema empty not right
+
+## 1.1.1-beta.0
+
+### Patch Changes
+
+- feat: native support sse and mcp serving by browser mcp
+
 ## 1.1.0
 
 ### Minor Changes